# feat(bitnet): implement Sub-Layer Normalization (SubLN) for #179