--- a/scripts/get-bitnet-pr-review-prompt.sh
+++ b/scripts/get-bitnet-pr-review-prompt.sh
@@ -3,7 +3,14 @@ TASK=$1
 PR=$2
 
 if test "x$TASK" = x; then
-  echo "USAGE: $0 TASK [PR]" >&2
+  TASK=$(./scripts/get-current-task-number.sh)
+fi
+if test "x$PR" = x; then
+  PR=$(./scripts/get-current-pr-number.sh)
+fi
+
+if test "x$TASK" = x || test "x$PR" = x; then
+  echo "USAGE: $0 [TASK [PR]]" >&2
   exit 0
 fi