--- a/scripts/get-bitnet-task-prompt.sh
+++ b/scripts/get-bitnet-task-prompt.sh
@@ -2,12 +2,15 @@
 TASK=$1
 PR=$2
 
+if test "x$TASK" = x; then
+  TASK=$(./scripts/get-current-task-number.sh)
+fi
 if test "x$PR" = x; then
-  PR=YOUR-PR-NUMBER
+  PR=$(./scripts/get-current-pr-number.sh)
 fi
 
-if test "x$TASK" = x; then
-  echo "USAGE: $0 TASK [PR]" >&2
+if test "x$TASK" = x || test "x$PR" = x; then
+  echo "USAGE: $0 [TASK [PR]]" >&2
   exit 0
 fi