--- a/scripts/get-bitnet-task-prompt.sh
+++ b/scripts/get-bitnet-task-prompt.sh
@@ -28,6 +28,15 @@ Focus strictly on GitHub issue #TASK#. That is the task. Do not touch unrelated
 files, do not refactor existing code, and do not fix things that aren't broken. 
 Extra changes mean extra review cycles and wasted time.
 
+```
+# Check current task info
+./scripts/get-current-task.sh|cat
+# Check current task number
+./scripts/get-current-task-number.sh|cat
+# Check current PR number
+./scripts/get-current-pr-number.sh|cat
+```
+
 The overall project direction is defined in GitHub issue #170. Keep that in 
 mind to avoid drifting off-course. To find all related issues, use the `bitnet` 
 and `task` labels in GitHub. These labels group all subtasks and planned work