## Test Coverage
- Current coverage: 85.0%
- Coverage changes: 84.1% → 85.0%

## Performance Metrics
### Memory Usage
#### Tensor Operations
- Allocations per operation:
  - New tensor creation: 120 allocs/op
  - Get/Set operations: 0 allocs/op
  - Parallel operations: 160749 allocs/op
  - BitLinear operations: 2101765 allocs/op

#### BitNet Model Operations
- Allocations per operation:
  - Model weights loading: 1289985000 allocs/op
  - Model inference: N/A (TODO #190) allocs/op (TODO #190)
  - Ternary weights reading: 48 allocs/op

### CPU Performance
#### Tensor Operations
- Operation timing:
  - Basic operations: 11.92 ns/op
  - Parallel operations: 95819 ns/op
  - Large tensor operations: 1282 ns/op
  - BitLinear operations: 1574935 ns/op

#### BitNet Model Operations
- Operation timing:
  - Model weights loading: 1349778167 ns/op
  - Model inference: BenchmarkModel_Infer ns/op (TODO #190)
  - Ternary weights reading: 3865 ns/op

## Areas for Improvement
### High Priority
- [ ] Optimize memory allocations in model operations (TODO #191)
- [ ] Implement proper self-attention (TODO #186)

### Medium Priority
- [ ] Improve error handling in model operations (TODO #192)
- [ ] Add more comprehensive benchmarks (TODO #192)
- [ ] Enhance documentation
- [ ] Implement proper feed-forward network (TODO #187)

### Low Priority
- [ ] Consider SIMD optimizations (TODO #191)
- [ ] Add more model operations (TODO #190)
- [ ] Improve test organization (TODO #192)
- [ ] Implement proper output generation (TODO #189)

Closes #179